// src/app/interfaces/salud.interface.ts
// Generated by https://quicktype.io

export interface ISalud {
    status:       string;
    totalResults: number;
    results:      Result[];
    nextPage:     string;
}

export interface Result {
    title:       string;
    link:        string;
    keywords:    string[] | null;
    creator:     string[] | null;
    video_url:   null;
    description: null | string;
    content:     string;
    pubDate:     string;
    image_url:   null | string;
    source_id:   SourceID;
    category:    Category[];
    country:     Country[];
    language:    Language;
    expanded?: boolean;
}

export enum Category {
    Top = "top",
}

export enum Country {
    Spain = "spain",
}

export enum Language {
    Spanish = "spanish",
}

export enum SourceID {
    Esinvesting = "esinvesting",
    Europapress = "europapress",
    HuffingtonpostEs = "huffingtonpost_es",
}